♻️ ::  [#1206] togglePopUp -> togglePopup 리네이밍
