+++ Projects/Features/MyInfoFeature/Sources/Components/SettingComponent.swift
@@ -17,7 +17,7 @@ public protocol SettingDependency: Dependency {
     var serviceTermsFactory: any ServiceTermFactory { get }
     var privacyFactory: any PrivacyFactory { get }
     var openSourceLicenseFactory: any OpenSourceLicenseFactory { get }
-    var togglePopUpFactory: any TogglePopUpFactory { get }
+    var togglePopupFactory: any TogglePopupFactory { get }
 }
 
 public final class SettingComponent: Component<SettingDependency>, SettingFactory {
@@ -33,7 +33,7 @@ public final class SettingComponent: Component<SettingDependency>, SettingFactor
             serviceTermsFactory: dependency.serviceTermsFactory,
             privacyFactory: dependency.privacyFactory,
             openSourceLicenseFactory: dependency.openSourceLicenseFactory,
-            togglePopUpFactory: dependency.togglePopUpFactory
+            togglePopupFactory: dependency.togglePopupFactory
         )
     }
 }
